separate tests for cache and storage
